feat: update implementation with suggested review comments
