+++ providers/rocketflag/README.md
@@ -15,4 +15,4 @@ go get github.com/open-feature/go-sdk-contrib/providers/rocketflag
 
 ## Usage
 
-Refer to the [sample.go](/providers/rocketflag/sample.go) file for how to use.
+Refer to the [sample.go](/providers/rocketflag/example_test.go) file for how to use.
